initial commit for error reporting in web worker
