+++ build-system/tasks/presubmit-checks.js
@@ -488,12 +488,12 @@ var forbiddenTerms = {
       'build-system/amp.extern.js',
       'build-system/tasks/prepend-global/test.js',
       'build-system/tasks/prepend-global/index.js',
-      'src/service-worker/core.js',
-      'src/service-worker/error-reporting.js',
-      'src/mode.js',
-      'src/experiments.js',
-      'src/config.js',
       'dist.3p/current/integration.js',
+      'src/service-worker/core.js',
+      'src/config.js',
+      'src/experiments.js',
+      'src/mode.js',
+      'src/worker-error-reporting.js',
     ],
   },
   'data:image/svg(?!\\+xml;charset=utf-8,)[^,]*,': {
